+++ kedro_mlflow/io/mlflow_dataset.py
@@ -1,4 +1,5 @@
 from typing import Any, Dict, Union
+from warnings import warn
 
 import mlflow
 from kedro.io import AbstractVersionedDataSet
@@ -6,7 +7,7 @@ from kedro.io.core import parse_dataset_definition
 from mlflow.tracking import MlflowClient
 
 
-class MlflowDataSet(AbstractVersionedDataSet):
+class MlflowArtifactDataSet(AbstractVersionedDataSet):
     """This class is a wrapper for any kedro AbstractDataSet.
     It decorates their ``save`` method to log the dataset in mlflow when ``save`` is called.
 
@@ -26,7 +27,7 @@ class MlflowDataSet(AbstractVersionedDataSet):
         # all dataset (i.e. it should replace AbstractVersionedDataSet)
         # instead and since we can't modify the core package,
         # we create a subclass which inherits dynamically from the data_set class
-        class MlflowDataSetChildren(data_set):
+        class MlflowArtifactDataSetChildren(data_set):
             def __init__(self, run_id, artifact_path):
                 super().__init__(**data_set_args)
                 self.run_id = run_id
@@ -56,31 +57,59 @@ class MlflowDataSet(AbstractVersionedDataSet):
 
         # rename the class
         parent_name = data_set.__name__
-        MlflowDataSetChildren.__name__ = f"Mlflow{parent_name}"
-        MlflowDataSetChildren.__qualname__ = f"{parent_name}.Mlflow{parent_name}"
+        MlflowArtifactDataSetChildren.__name__ = f"Mlflow{parent_name}"
+        MlflowArtifactDataSetChildren.__qualname__ = (
+            f"{parent_name}.Mlflow{parent_name}"
+        )
 
-        mlflow_dataset_instance = MlflowDataSetChildren(
+        mlflow_dataset_instance = MlflowArtifactDataSetChildren(
             run_id=run_id, artifact_path=artifact_path
         )
         return mlflow_dataset_instance
 
     def _load(self) -> Any:  # pragma: no cover
         """
-        MlowDataSet is a factory for DataSet
+        MlflowArtifactDataSet is a factory for DataSet
         and consequently does not implements abtracts methods
         """
         pass
 
     def _save(self, data: Any) -> None:  # pragma: no cover
         """
-        MlowDataSet is a factory for DataSet
+        MlflowArtifactDataSet is a factory for DataSet
         and consequently does not implements abtracts methods
         """
         pass
 
     def _describe(self) -> Dict[str, Any]:  # pragma: no cover
         """
-        MlowDataSet is a factory for DataSet
+        MlflowArtifactDataSet is a factory for DataSet
         and consequently does not implements abtracts methods
         """
         pass
+
+
+class MlflowDataSet(MlflowArtifactDataSet):
+    def __new__(
+        cls,
+        data_set: Union[str, Dict],
+        run_id: str = None,
+        artifact_path: str = None,
+        credentials: Dict[str, Any] = None,
+    ):
+        deprecation_msg = (
+            "'MlflowDataSet' is now deprecated and "
+            "has been renamed 'MlflowArtifactDataSet' "
+            "in 'kedro-mlflow>=0.3.0'. "
+            "\nPlease change your 'catalog.yml' entries accordingly, "
+            "since 'MlflowDataSet' will be removed in next release."
+        )
+
+        warn(deprecation_msg, DeprecationWarning, stacklevel=2)
+        super().__new__(
+            cls=cls,
+            data_set=data_set,
+            run_id=run_id,
+            artifact_path=artifact_path,
+            credentials=credentials,
+        )